Rename test files.
