+++ packages/file_selector/file_selector_windows/test/fake_shell_item.dart
